requirements.txt
streamlit
google.generativeai
base64
requests
Image
langchain